Retriever_Agent/requirements.txt
# Core LangChain + RAG support
langchain
langchain-community
langchain-chroma
langchain-groq

# Vector DB
chromadb

# Hugging Face embeddings
langchain-huggingface
sentence-transformers

# Database access
sqlalchemy
psycopg2-binary

# UI
gradio

# Utilities
python-dotenv